Move the hiredis_libevent2 code from common to relay
