+++ src/apps/common/CMakeLists.txt
@@ -52,15 +52,6 @@ else()
         message(FATAL_ERROR "Must set Libevent")
     endif()
 endif()
-
-find_package(hiredis)
-if(hiredis_FOUND)
-    list(APPEND SOURCE_FILES hiredis_libevent2.c)
-    list(APPEND HEADER_FILES hiredis_libevent2.h)
-    list(APPEND COMMON_LIBS hiredis::hiredis)
-else()
-    list(APPEND COMMON_DEFINED TURN_NO_HIREDIS)
-endif()
 message("COMMON_LIBS:${COMMON_LIBS}")
 
 add_library(${PROJECT_NAME} STATIC ${SOURCE_FILES} ${HEADER_FILES})
