+++ src/apps/relay/CMakeLists.txt
@@ -98,8 +98,8 @@ endif()
 find_package(hiredis)
 if(hiredis_FOUND)
     list(APPEND turnserver_LIBS hiredis::hiredis)
-    list(APPEND SOURCE_FILES dbdrivers/dbd_redis.c)
-    list(APPEND HEADER_FILES dbdrivers/dbd_redis.h)
+    list(APPEND SOURCE_FILES hiredis_libevent2.c dbdrivers/dbd_redis.c)
+    list(APPEND HEADER_FILES hiredis_libevent2.h dbdrivers/dbd_redis.h)
 else()
     list(APPEND turnserver_DEFINED TURN_NO_HIREDIS)
 endif()
